Full move of MathJax to the UI component
